abstract functions
